+++ src/main/java/com/clone/team4/domain/like/repository/LikeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.clone.team4.domain.like.entity.Like;
 
-public interface LikeRepository extends JpaRepository<Like, Long> {
+public interface LikeRepository extends JpaRepository<Like, Long>, QLikeRepository {
 }
